Reason for Change: split NetworkBandwidthEstimator to separate class, removing libcurl dependencies
